fix(reconciler): remove hardcoded sslmode config
previously, ssl mode was not configurable when creating a connection to
the postgres database. while the lack of SSL is acceptable for local
deployments (i.e., dockerized postgres in-cluster), we hit security
restrictions when trying to connect to external databases (like AWS).
this commit allows the operator to configure the ssl mode to their
liking.